+++ lys/widgets/canvas/Matplotlib/AreaSettings.py
@@ -47,4 +47,4 @@ class _MatplotlibCanvasSize(CanvasSize):
         m = self.canvas().getMargin()
         for ax in self.canvas().axisList():
             self.canvas().getAxes(ax).set_position([m[0], m[2], m[1] - m[0], m[3] - m[2]])
-        self.canvas().resize(self.canvas().getFigure().get_figwidth() * 100, self.canvas().getFigure().get_figheight() * 100)
+        self.canvas().resize(int(self.canvas().getFigure().get_figwidth() * 100), int(self.canvas().getFigure().get_figheight() * 100))
